Refactor into separate `customLinter` file
